chore(audit): batch D — 14 medium/low follow-ups
Final batch of low-risk audit follow-ups for the 2026-04-26 audit. All
on the same `chore/audit-fixes-2026-04-26` branch.

## Backend / Python (Medium)

- **M3** `api/cache.py` — replace unbounded `_locks: dict` with
  `cachetools.LRUCache(maxsize=cache_maxsize * 2)`. The lock dict was
  growing unbounded across long uptime (~hundreds of `filter:lib=...`
  and `og:{spec}:{lang}:{lib}` variants); `clear_cache_by_pattern`
  cleared `_cache` and `_timestamps` but not `_locks`. LRU eviction
  fixes that without changing the dict-like interface.
- **M5** `core/images.py` — `subprocess.run(["pngquant","--version"])`
  at module import time → `timeout=5`; `optimize_png` subprocess →
  `timeout=60`; catch `TimeoutExpired`. A wedged binary can no longer
  block FastAPI startup.
- **M6** `api/routers/download.py` — `httpx.AsyncClient(timeout=15.0)`
  (was unbounded). Other handlers in this router tree already use
  10–30 s; 15 s for image downloads is consistent.

## Frontend (Medium)

- **M10** `ErrorBoundary.tsx` + `RouteErrorBoundary.tsx` — replace MUI
  default tokens (`color="text.secondary"`, `bgcolor: 'grey.100'`)
  with project CSS vars (`var(--ink-muted)`, `var(--bg-surface)`,
  `var(--ink-soft)`). Fixes low-contrast fallback in dark mode —
  exactly when the error is shown.
- **M11** Same two files — convert MUI barrel imports
  (`from '@mui/material'`) to per-component subpath imports. Matches
  the convention used in 50+ other files; helps tree-shaking.
- **M12** New `app/src/utils/shuffle.ts` exporting `shuffleArray`
  (Fisher-Yates). `useFilterFetch.ts` now imports it (drop the local
  copy); `useFeaturedSpecs.ts` switches from the biased
  `[...arr].sort(() => Math.random() - 0.5)` idiom to the uniform
  shuffle.
- **M14** `FilterBar.tsx`, `PlotsPage.tsx`, `SpecsListPage.tsx` —
  scroll listeners now use `{ passive: true }` so the browser doesn't
  wait for our handler to ack before scrolling. Matters most on
  `/plots`, the busiest scroll path.
- **L12** `SpecOverview.tsx` — wrap the alphabetical sort in
  `useMemo([implementations])` so it doesn't run on every parent
  re-render (tooltip toggle, copy-state changes, etc.).

## Backend low-risk (Low)

- **L1** `api/routers/insights.py` and `core/database/repositories.py`
  — move `import random` and `from sqlalchemy import String, cast,
  or_` from inside function bodies to module top.
- **L21** `api/analytics.py:124` — promote silent
  `logger.debug("Plausible tracking failed …")` to `logger.warning`.
  Cloud Run typically suppresses DEBUG, which made full-pipeline
  Plausible outages invisible in incident triage.

## Cleanup (Low)

- **L7** `rm -rf automation/generators/` — stale `__pycache__` from
  the deleted `core.generators.plot_generator` module (Batch A
  deleted the source; this removes the leftover bytecode).
- **L8** `rm -rf tests/unit/plots/` — empty test package with only
  `__init__.py "Unit tests for plot implementations."` but no tests
  (plot validation is pipeline-driven via `impl-review.yml`).
- **L9** `docs/reference/{api,database,repository}.md` +
  `agentic/docs/project-guide.md` — sed
  `claude-opus-4-5-20251101` → `claude-opus-4-7-20260101` in
  `generated_by` examples.
- **L22** `docs/reference/plausible.md` — remove `signal` from the
  `platform` values list and add `whatsapp-lite` (the actual value
  emitted by `api/analytics.py:81` for Signal-spoofed UAs); add a
  short note explaining the alias.

## Skipped from the original list of 16

- **M23** workflow `defaults: { run: { shell: 'bash -euxo pipefail
  {0}' } }` — GitHub already runs `bash -eo pipefail` by default.
  Adding `-u` (nounset) would break legit `[ -z "$VAR" ]` patterns;
  `-x` (xtrace) is too noisy for prod logs. No-op or risky — skipped.
- **L11** `useThemeMode.ts` — the audit's SSR guard suggestion is
  already in place (`typeof window !== 'undefined'` in
  `systemPrefersDark()`); the `setMode` / `cycle` "dedup" would be
  cosmetic. Skipped.

## Verification

- `uv run ruff check .` — passes
- `uv run ruff format --check .` — clean (118 files)
- `uv run pytest tests/unit` — 1312 passed
- `cd app && yarn tsc --noEmit` — clean
- `cd app && yarn test --run` on SpecOverview / useFilterFetch /
  useFeaturedSpecs — 22 tests pass
