+++ app/src/pages/PlotsPage.tsx
@@ -113,7 +113,8 @@ export function PlotsPage() {
     const handleScroll = () => {
       setShowScrollTop(window.scrollY > 300);
     };
-    window.addEventListener('scroll', handleScroll);
+    // passive: true — handler doesn't preventDefault.
+    window.addEventListener('scroll', handleScroll, { passive: true });
     return () => window.removeEventListener('scroll', handleScroll);
   }, []);
