+++ app/src/pages/SpecsListPage.tsx
@@ -111,7 +111,8 @@ export function SpecsListPage() {
     const handleScroll = () => {
       setShowScrollTop(window.scrollY > 300);
     };
-    window.addEventListener('scroll', handleScroll);
+    // passive: true — handler doesn't preventDefault.
+    window.addEventListener('scroll', handleScroll, { passive: true });
     return () => window.removeEventListener('scroll', handleScroll);
   }, []);
