--- tests/unit/plots/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for plot implementations."""
